Guess nothing is really easy with intellij sdk. Dispose the service after each run to make sure we have clean state after each run
